+++ csaf-cvss/src/main/kotlin/io/github/csaf/sbom/cvss/Calculation.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2024, The Authors. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+package io.github.csaf.sbom.cvss
+
+/**
+ * Represents a CVSS score calculation. This needs to be implemented by the respective standard
+ * versions.
+ */
+interface CvssCalculation {
+
+    val metrics: Map<String, String>
+
+    fun calculateBaseScore(): Double
+
+    fun calculateTemporalScore(): Double
+
+    fun calculateEnvironmentalScore(): Double
+}
